Refactorisation du composant footer, correction de l'ensemble des imports liée a ce composant
