+++ client/src/pages/ComptePage/ComptePage.tsx
@@ -1,7 +1,7 @@
 import style from "../stylePage.module.css";
 import NavRoot from "../../components/Layout/NavRoot/NavRoot";
 import CompteRoot from "../../components/CompteRoot/CompteRoot";
-import FooterRoot from "../../components/FooterRoot/FooterRoot";
+import FooterRoot from "../../components/Layout/FooterRoot/FooterRoot";
 
 function ComptePage() {
     return (
