+++ client/src/pages/Public/HomePage/HomePage.tsx
@@ -1,7 +1,7 @@
 import css from "../../stylePage.module.css"
 import NavTargetHomeRoot from "../../../components/NavTargetHomeRoot/NavTargetHomeRoot";
 import HomeRoot from "../../../components/Public/HomeRoot/HomeRoot";
-import FooterRoot from "../../../components/FooterRoot/FooterRoot";
+import FooterRoot from "../../../components/Layout/FooterRoot/FooterRoot";
 import Helmet_Home_SEO from "../../../seo/HomeSEO/Helmet_Home.seo";
 
 function HomePage() {
