+++ client/src/pages/Public/MentionsLegalesPage/MentionsLegalesPage.tsx
@@ -1,7 +1,7 @@
 import style from "../../stylePage.module.css";
 import NavRoot from "../../../components/Layout/NavRoot/NavRoot";
 import MentionsLegalesRoot from "../../../components/Public/MentionsLegalesRoot/MentionsLegalesRoot";
-import FooterRoot from "../../../components/FooterRoot/FooterRoot";
+import FooterRoot from "../../../components/Layout/FooterRoot/FooterRoot";
 import Helmet_MentionLegale_SEO from "../../../seo/MentionsLegalesSEO/Helmet_MentionsLegales.seo";
 import useScrollToHash from "../../../hook/useScrollToHash";
