+++ client/src/pages/Public/PortraitCorporatePage/PortraitCorporatePage.tsx
@@ -1,7 +1,7 @@
 import style from "../../stylePage.module.css";
 import NavRoot from "../../../components/Layout/NavRoot/NavRoot";
 import PortraitCorporateRoot from "../../../components/Public/PortraitCorporateRoot/PortraitCorporateRoot";
-import FooterRoot from "../../../components/FooterRoot/FooterRoot";
+import FooterRoot from "../../../components/Layout/FooterRoot/FooterRoot";
 import Helmet_PortraitCorporate_SEO from "../../../seo/PortraitCorporateSEO/Helmet_PortraitCorporate.seo";
 import useScrollToHash from "../../../hook/useScrollToHash";
