+++ client/src/pages/Public/PortraitDuoPage/PortraitDuoPage.tsx
@@ -1,7 +1,7 @@
 import style from "../../stylePage.module.css";
 import NavRoot from "../../../components/Layout/NavRoot/NavRoot";
 import PortraitDuoRoot from "../../../components/Public/PortraitDuoRoot/PortraitDuoRoot";
-import FooterRoot from "../../../components/FooterRoot/FooterRoot";
+import FooterRoot from "../../../components/Layout/FooterRoot/FooterRoot";
 import Helmet_PortraitDuo_SEO from "../../../seo/PortraitDuoSEO/Helmet_PortraitDuo.seo";
 import useScrollToHash from "../../../hook/useScrollToHash";
