+++ client/src/pages/Public/PortraitFamillePage/PortraitFamillePage.tsx
@@ -1,7 +1,7 @@
 import style from "../../stylePage.module.css";
 import NavRoot from "../../../components/Layout/NavRoot/NavRoot";
 import PortraitFamilleRoot from "../../../components/Public/PortraitFamilleRoot/PortraitFamilleRoot";
-import FooterRoot from "../../../components/FooterRoot/FooterRoot";
+import FooterRoot from "../../../components/Layout/FooterRoot/FooterRoot";
 import Helmet_PortraitFamille_SEO from "../../../seo/PortraitFamilleSEO/Helmet_PortraitFamille.seo";
 import useScrollToHash from "../../../hook/useScrollToHash";
