+++ client/src/pages/Public/PortraitLifeStylePage/PortraitLifeStylePage.tsx
@@ -1,7 +1,7 @@
 import NavRoot from "../../../components/Layout/NavRoot/NavRoot";
 import style from "../../stylePage.module.css";
 import PortraitLifeStyleRoot from "../../../components/Public/PortraitLifeStyleRoot/PortraitLifeStyleRoot";
-import FooterRoot from "../../../components/FooterRoot/FooterRoot";
+import FooterRoot from "../../../components/Layout/FooterRoot/FooterRoot";
 import Helmet_PortraitLifeStyle_SEO from "../../../seo/PortraitLifeStyleSEO/Helmet_PortraitLifeStyle.seo";
 
 function PortraitLifeStylePage() {
