+++ client/src/pages/Public/PortraitMariagePage/PortraitMariagePage.tsx
@@ -1,7 +1,7 @@
 import style from "../../stylePage.module.css";
 import NavRoot from "../../../components/Layout/NavRoot/NavRoot";
 import PortraitMariageRoot from "../../../components/Public/PortraitMariageRoot/MariageRoot";
-import FooterRoot from "../../../components/FooterRoot/FooterRoot";
+import FooterRoot from "../../../components/Layout/FooterRoot/FooterRoot";
 import Helmet_PortraitMariage_SEO from "../../../seo/PortraitMariageSEO/Helmet_PortraitMariage.seo";
 import useScrollToHash from "../../../hook/useScrollToHash";
