+++ client/src/pages/Public/PrestationPage/PrestationPage.tsx
@@ -1,7 +1,7 @@
 import style from "../../stylePage.module.css";
 import NavRoot from "../../../components/Layout/NavRoot/NavRoot";
 import PrestationRoot from "../../../components/Public/PrestationRoot/PrestationRoot";
-import FooterRoot from "../../../components/FooterRoot/FooterRoot";
+import FooterRoot from "../../../components/Layout/FooterRoot/FooterRoot";
 import Helmet_Prestation_SEO from "../../../seo/PrestationSEO/Helmet_Prestation.seo";
 
 function TarifsPage() {
